+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Accept code suggestions line by line ([!200](https://gitlab.com/gitlab-org/editor-extensions/gitlab-eclipse-plugin/-/merge_requests/200))
+
 ### Changed
 
 - Changed the log level field in the preferences page from a text field into a dropdown. ([!216](https://gitlab.com/gitlab-org/editor-extensions/gitlab-eclipse-plugin/-/merge_requests/216))